+++ ShyamGui/Source/MeasurementData.h
@@ -1325,7 +1325,7 @@ namespace MeasurementData
         };
 
         line ("SourceDesc=Atomik_Data_Text");
-        line ("Version='v1.4.0.5'");
+        line ("Version='v1.4.0.6'");
         line ("Author='Atomik Prediction Software'");
         line ("");
         line ("Frequency_Hz=" + juce::String (hz));